refactor: 도메인 변경 및 조회 메서드 수정
도메인을 Contents로 옮기고 조회 메서드를 findById 메서드로 변경하였습니다.

+++ src/main/java/com/example/tripKo/domain/place/dao/PlaceRestaurantJPARepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PlaceRestaurantJPARepository extends JpaRepository<PlaceRestaurant, Integer> {
+public interface PlaceRestaurantJPARepository extends JpaRepository<PlaceRestaurant, Long> {
     @Query("SELECT p FROM PlaceRestaurant p WHERE UPPER(p.place.address.addressCategory.sidoName) = UPPER(:location)")
     List<PlaceRestaurant> findAllByLocation(@Param("location") String location);
 
@@ -28,5 +28,6 @@ public interface PlaceRestaurantJPARepository extends JpaRepository<PlaceRestaur
 //            "LEFT JOIN FETCH c.contentsMenus m " +
 //            "LEFT JOIN FETCH c.contentsFiles cf " +
             "WHERE r.id = :id")
-    PlaceRestaurant findRestaurantDetailsById(@Param("id")long id);
+    PlaceRestaurant findRestaurantDetailsById(@Param("id")Long id);
+
 }
